Optimisdations

+++ src/Logger.php
@@ -9,12 +9,21 @@ class Logger extends Singleton
     public function writeLine(string $line): void
     {
         $this->ensureLogFileExists();
-        // add header if missing
         if (empty($this->lines)) {
-            $line = $this->getHeaderLine();
-            $key = md5($line);
-            file_put_contents($this->getPath(), $line . "\n");
-            $this->lines[$key] = true;
+            $existingLogFile = file_get_contents($this->getPath());
+            if (empty($existingLogFile)) {
+                // add header if missing
+                $line = $this->getHeaderLine();
+                $key = md5($line);
+                file_put_contents($this->getPath(), $line . "\n");
+                $this->lines[$key] = true;
+            } else {
+                // hydrate $this->lines with existing contents of log file
+                foreach (explode("\n", $existingLogFile) as $line) {
+                    $key = md5($line);
+                    $this->lines[$key] = true;
+                }
+            }
         }
         // only write unique lines
         $key = md5($line);
@@ -24,12 +33,6 @@ class Logger extends Singleton
         $line = trim($line, "\n") . "\n";
         file_put_contents($this->getPath(), $line, FILE_APPEND);
         $this->lines[$key] = true;
-    }
-
-    public function clearLogFile(): void
-    {
-        $this->ensureLogFileExists();
-        file_put_contents($this->getPath(), '');
     }
 
     private function ensureLogFileExists(): void
@@ -43,6 +46,12 @@ class Logger extends Singleton
             mkdir($dir);
         }
         file_put_contents($path, '');
+    }
+
+    public function clearLogFile(): void
+    {
+        $this->ensureLogFileExists();
+        file_put_contents($this->getPath(), '');
     }
 
     private function getHeaderLine(): string
